+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+### 3.6.2
+  * Set max_size of pagination to 30 for fullfilment_orders [#230](https://github.com/singer-io/tap-shopify/pull/230)
+
+## 3.6.1
+  * Dependency upgrades [#228](https://github.com/singer-io/tap-shopify/pull/228)
+
+## 3.6.0
+  * Introduce new stream fulfillment_orders. [#222](https://github.com/singer-io/tap-shopify/pull/222)
+
+## 3.5.0
+  * Add fields in orders stream - retailLocation and location ID (within fulfillments) [#219](https://github.com/singer-io/tap-shopify/pull/219)
+  * Skip the author field on missing read_users scope [#223](https://github.com/singer-io/tap-shopify/pull/223)
+
 ## 3.4.0
   * Introduce new stream order_shipping_lines. [#213](https://github.com/singer-io/tap-shopify/pull/213)
   * shippingLine is a nested object within the orders stream and it represents individual shipping methods.